basic review feed done

+++ src/main/java/com/google/codeu/data/Review.java
@@ -7,7 +7,7 @@ public class Review {
   private final UUID reviewId; //This is the ID of the review
   private long timestamp;
   private String author; //The user posting the review
-  private int rating;
+  private long rating;
   private String comment;
   private List<String> pictures;
 
@@ -26,7 +26,7 @@ public class Review {
    * Creates a new Review object for every new review for cases when
    * the user wants to provide a complete review.
    */
-  public Review(String author, int rating, String comment,
+  public Review(String author, long rating, String comment,
                 List<String> pictures) {
     this.reviewId = UUID.randomUUID();
     this.author = author;
@@ -36,7 +36,7 @@ public class Review {
     this.timestamp = System.currentTimeMillis();
   }
 
-  public Review(String author, int rating, String comment, long timestamp) {
+  public Review(String author, long rating, String comment, long timestamp) {
     this.reviewId = UUID.randomUUID();
     this.author = author;
     this.rating = rating;
@@ -61,7 +61,7 @@ public class Review {
     this.author = author;
   }
 
-  public int getRating() {
+  public long getRating() {
     return rating;
   }
